requirements.txt
python-dotenv==1.0.1
nest-asyncio==1.6.0
discord.py==2.4.0
cuml-cu11==24.12.0
cudf-cu11==24.12.0
seaborn==0.13.2
matplotlib==3.10.0
transformers==4.48.3
scikit-learn==1.6.1
numpy==2.0.2
pandas==2.2.3
torch==2.6.0+cu118